+++ picocli-annotation-processing-tests/src/test/java/picocli/annotation/processing/tests/Issue1316Test.java
@@ -1,0 +1,24 @@
+package picocli.annotation.processing.tests;
+
+import com.google.testing.compile.Compilation;
+import com.google.testing.compile.JavaFileObjects;
+import org.junit.Test;
+
+import javax.annotation.processing.Processor;
+
+import static com.google.testing.compile.CompilationSubject.assertThat;
+import static com.google.testing.compile.Compiler.javac;
+
+public class Issue1316Test {
+    @Test
+    public void testIssue1316() {
+        Processor processor = new AnnotatedCommandSourceGeneratorProcessor();
+        Compilation compilation =
+                javac()
+                        .withProcessors(processor)
+                        .compile(JavaFileObjects.forResource(
+                                "picocli/issue1316/Example.java"));
+
+        assertThat(compilation).succeeded();
+    }
+}